backend/requirements.txt
Django==4.2.7
mysqlclient==2.2.0
PyMySQL==1.1.2 
Pillow==10.1.0
python-decouple==3.8
openpyxl==3.1.2
cryptography==46.0.3
django-cors-headers==4.9.0
gunicorn==21.2.0

